+++ example/lib/pages/card.dart
@@ -98,7 +98,7 @@ class _CardPageState extends State<CardPage> {
             width: double.infinity,
             icon: const Padding(
               padding: EdgeInsets.only(right: 8),
-              child: ShadImage.square(LucideIcons.check, size: 16),
+              child: Icon(LucideIcons.check),
             ),
             onPressed: () {},
             child: const Text('Mark all as read'),
@@ -114,7 +114,7 @@ class _CardPageState extends State<CardPage> {
                 ),
                 child: Row(
                   children: [
-                    ShadImage.square(
+                    Icon(
                       LucideIcons.bellRing,
                       size: 24,
                       color: theme.colorScheme.foreground,