+++ lib/src/components/button.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:shadcn_ui/src/components/image.dart';
 import 'package:shadcn_ui/src/raw_components/focusable.dart';
 import 'package:shadcn_ui/src/theme/components/button.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
@@ -10,7 +9,6 @@ import 'package:shadcn_ui/src/theme/theme.dart';
 import 'package:shadcn_ui/src/utils/debug_check.dart';
 import 'package:shadcn_ui/src/utils/extensions/order_policy.dart';
 import 'package:shadcn_ui/src/utils/gesture_detector.dart';
-import 'package:shadcn_ui/src/utils/provider.dart';
 import 'package:shadcn_ui/src/utils/separated_iterable.dart';
 import 'package:shadcn_ui/src/utils/states_controller.dart';
 
@@ -36,7 +34,6 @@ class ShadButton extends StatefulWidget {
     this.icon,
     this.onPressed,
     this.size,
-    this.applyIconColorFilter,
     this.cursor,
     this.width,
     this.height,
@@ -81,7 +78,6 @@ class ShadButton extends StatefulWidget {
     this.onFocusChange,
     this.orderPolicy,
     this.expands,
-    this.iconSize,
   }) : variant = ShadButtonVariant.primary;
 
   const ShadButton.raw({
@@ -91,7 +87,6 @@ class ShadButton extends StatefulWidget {
     this.child,
     this.icon,
     this.onPressed,
-    this.applyIconColorFilter,
     this.cursor,
     this.width,
     this.height,
@@ -136,7 +131,6 @@ class ShadButton extends StatefulWidget {
     this.onFocusChange,
     this.orderPolicy,
     this.expands,
-    this.iconSize,
   });
 
   const ShadButton.destructive({
@@ -145,7 +139,6 @@ class ShadButton extends StatefulWidget {
     this.icon,
     this.onPressed,
     this.size,
-    this.applyIconColorFilter,
     this.cursor,
     this.width,
     this.height,
@@ -190,7 +183,6 @@ class ShadButton extends StatefulWidget {
     this.onFocusChange,
     this.orderPolicy,
     this.expands,
-    this.iconSize,
   }) : variant = ShadButtonVariant.destructive;
 
   const ShadButton.outline({
@@ -199,7 +191,6 @@ class ShadButton extends StatefulWidget {
     this.icon,
     this.onPressed,
     this.size,
-    this.applyIconColorFilter,
     this.cursor,
     this.width,
     this.height,
@@ -244,7 +235,6 @@ class ShadButton extends StatefulWidget {
     this.onFocusChange,
     this.orderPolicy,
     this.expands,
-    this.iconSize,
   }) : variant = ShadButtonVariant.outline;
 
   const ShadButton.secondary({
@@ -253,7 +243,6 @@ class ShadButton extends StatefulWidget {
     this.icon,
     this.onPressed,
     this.size,
-    this.applyIconColorFilter,
     this.cursor,
     this.width,
     this.height,
@@ -298,7 +287,6 @@ class ShadButton extends StatefulWidget {
     this.onFocusChange,
     this.orderPolicy,
     this.expands,
-    this.iconSize,
   }) : variant = ShadButtonVariant.secondary;
 
   const ShadButton.ghost({
@@ -307,7 +295,6 @@ class ShadButton extends StatefulWidget {
     this.icon,
     this.onPressed,
     this.size,
-    this.applyIconColorFilter,
     this.cursor,
     this.width,
     this.height,
@@ -352,7 +339,6 @@ class ShadButton extends StatefulWidget {
     this.onFocusChange,
     this.orderPolicy,
     this.expands,
-    this.iconSize,
   }) : variant = ShadButtonVariant.ghost;
 
   const ShadButton.link({
@@ -360,7 +346,6 @@ class ShadButton extends StatefulWidget {
     required this.child,
     this.onPressed,
     this.size,
-    this.applyIconColorFilter,
     this.cursor,
     this.width,
     this.height,
@@ -406,8 +391,7 @@ class ShadButton extends StatefulWidget {
     this.orderPolicy,
     this.expands,
   })  : variant = ShadButtonVariant.link,
-        icon = null,
-        iconSize = null;
+        icon = null;
 
   final VoidCallback? onPressed;
   final VoidCallback? onLongPress;
@@ -415,7 +399,7 @@ class ShadButton extends StatefulWidget {
   final Widget? child;
   final ShadButtonVariant variant;
   final ShadButtonSize? size;
-  final bool? applyIconColorFilter;
+
   final MouseCursor? cursor;
   final double? width;
   final double? height;
@@ -487,11 +471,6 @@ class ShadButton extends StatefulWidget {
   /// Whether the child expands to fill the available space, defaults to false.
   /// {@endtemplate}
   final bool? expands;
-
-  /// {@template ShadButton.iconSize}
-  /// The size of the icon, defaults to `Size.square(16)`
-  /// {@endtemplate}
-  final Size? iconSize;
 
   @override
   State<ShadButton> createState() => _ShadButtonState();
@@ -715,14 +694,12 @@ class _ShadButtonState extends State<ShadButton> {
     assertCheckHasTextOrIcon();
 
     final theme = ShadTheme.of(context);
+    final materialTheme = Theme.of(context);
 
     final hasPressedBackgroundColor = widget.pressedBackgroundColor != null ||
         buttonTheme(theme).pressedBackgroundColor != null;
     final hasPressedForegroundColor = widget.pressedForegroundColor != null ||
         buttonTheme(theme).pressedForegroundColor != null;
-
-    final applyIconColorFilter = widget.applyIconColorFilter ??
-        theme.primaryButtonTheme.applyIconColorFilter;
 
     final effectiveDecoration =
         (buttonTheme(theme).decoration ?? const ShadDecoration())
@@ -755,9 +732,6 @@ class _ShadButtonState extends State<ShadButton> {
     final effectiveExpands =
         widget.expands ?? buttonTheme(theme).expands ?? false;
 
-    final effectiveIconSize =
-        widget.iconSize ?? buttonTheme(theme).iconSize ?? const Size.square(16);
-
     return CallbackShortcuts(
       bindings: {
         const SingleActivator(LogicalKeyboardKey.enter): onTap,
@@ -769,49 +743,29 @@ class _ShadButtonState extends State<ShadButton> {
           final hovered = states.contains(ShadState.hovered);
           final enabled = !states.contains(ShadState.disabled);
 
+          final effectiveBackgroundColor = hasPressedBackgroundColor && pressed
+              ? pressedBackgroundColor(theme)
+              : hovered
+                  ? hoverBackground(theme)
+                  : background(theme);
+
+          final effectiveForegroundColor = hasPressedForegroundColor && pressed
+              ? pressedForegroundColor(theme)
+              : hovered
+                  ? hoverForeground(theme)
+                  : foreground(theme);
+
           final updatedDecoration = effectiveDecoration.copyWith(
-            color: hasPressedBackgroundColor && pressed
-                ? pressedBackgroundColor(theme)
-                : hovered
-                    ? hoverBackground(theme)
-                    : background(theme),
+            color: effectiveBackgroundColor,
             gradient: gradient(theme),
             shadows: shadows(theme),
           );
-
-          var icon = widget.icon;
-
-          // Applies the foreground color filter to the icon if provided
-          if (icon != null && (applyIconColorFilter ?? true)) {
-            icon = ColorFiltered(
-              colorFilter: ColorFilter.mode(
-                hasPressedForegroundColor && pressed
-                    ? pressedForegroundColor(theme)
-                    : hovered
-                        ? hoverForeground(theme)
-                        : foreground(theme),
-                BlendMode.srcIn,
-              ),
-              child: SizedBox.fromSize(
-                size: effectiveIconSize,
-                child: ShadProvider<ShadImageSize>(
-                  data: ShadImageSize.copy(effectiveIconSize),
-                  notifyUpdate: (state) => effectiveIconSize != state.data,
-                  child: icon,
-                ),
-              ),
-            );
-          }
 
           Widget? child = widget.child == null
               ? null
               : DefaultTextStyle(
                   style: theme.textTheme.small.copyWith(
-                    color: hasPressedForegroundColor && pressed
-                        ? pressedForegroundColor(theme)
-                        : hovered
-                            ? hoverForeground(theme)
-                            : foreground(theme),
+                    color: effectiveForegroundColor,
                     decoration: textDecoration(
                       theme,
                       hovered: hovered,
@@ -897,7 +851,15 @@ class _ShadButtonState extends State<ShadButton> {
                           mainAxisAlignment: effectiveMainAxisAlignment,
                           textDirection: effectiveTextDirection,
                           children: [
-                            if (icon != null) icon,
+                            if (widget.icon != null)
+                              Theme(
+                                data: materialTheme.copyWith(
+                                  iconTheme: materialTheme.iconTheme.copyWith(
+                                    color: effectiveForegroundColor,
+                                  ),
+                                ),
+                                child: widget.icon!,
+                              ),
                             if (child != null) child,
                           ]
                               .order(effectiveOrderPolicy)