+++ lib/src/components/checkbox.dart
@@ -3,7 +3,6 @@ import 'package:flutter/services.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 import 'package:lucide_icons_flutter/lucide_icons.dart';
 import 'package:shadcn_ui/src/components/disabled.dart';
-import 'package:shadcn_ui/src/components/image.dart';
 import 'package:shadcn_ui/src/raw_components/focusable.dart';
 import 'package:shadcn_ui/src/theme/components/decorator.dart';
 import 'package:shadcn_ui/src/theme/theme.dart';
@@ -130,7 +129,7 @@ class _ShadCheckboxState extends State<ShadCheckbox> {
     final effectiveSize = widget.size ?? theme.checkboxTheme.size ?? 16;
 
     final effectiveIcon = widget.icon ??
-        ShadImage.square(
+        Icon(
           LucideIcons.check,
           color: theme.colorScheme.primaryForeground,
           size: effectiveSize,