+++ lib/src/components/select.dart
@@ -7,7 +7,6 @@ import 'package:flutter/services.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 import 'package:lucide_icons_flutter/lucide_icons.dart';
 import 'package:shadcn_ui/src/components/disabled.dart';
-import 'package:shadcn_ui/src/components/image.dart';
 import 'package:shadcn_ui/src/components/input.dart';
 import 'package:shadcn_ui/src/components/popover.dart';
 import 'package:shadcn_ui/src/raw_components/focusable.dart';
@@ -656,7 +655,7 @@ class ShadSelectState<T> extends State<ShadSelect<T>> {
     }
 
     final effectiveTrailing = widget.trailing ??
-        ShadImage.square(
+        Icon(
           LucideIcons.chevronDown,
           size: 16,
           color: theme.colorScheme.popoverForeground.withValues(alpha: .5),
@@ -805,7 +804,7 @@ class ShadSelectState<T> extends State<ShadSelect<T>> {
                                   width: calculatedMinWidth,
                                   padding:
                                       const EdgeInsets.symmetric(vertical: 4),
-                                  child: ShadImage.square(
+                                  child: Icon(
                                     LucideIcons.chevronUp,
                                     size: 16,
                                     color: theme.colorScheme.popoverForeground,
@@ -835,7 +834,7 @@ class ShadSelectState<T> extends State<ShadSelect<T>> {
                                   width: calculatedMinWidth,
                                   padding:
                                       const EdgeInsets.symmetric(vertical: 4),
-                                  child: ShadImage.square(
+                                  child: Icon(
                                     LucideIcons.chevronDown,
                                     size: 16,
                                     color: theme.colorScheme.popoverForeground,
@@ -1035,7 +1034,7 @@ class _ShadOptionState<T> extends State<ShadOption<T>> {
           visible: selected,
           child: Padding(
             padding: const EdgeInsets.only(right: 8),
-            child: ShadImage.square(
+            child: Icon(
               LucideIcons.check,
               size: 16,
               color: theme.colorScheme.popoverForeground,