+++ playground/lib/pages/button.dart
@@ -59,7 +59,13 @@ class ButtonPage extends StatelessWidget {
             ),
           PlagroundButtonStyle.loading => ShadButton(
               onPressed: () {},
-              icon: const CircularProgressIndicator(strokeWidth: 2),
+              icon: SizedBox.square(
+                dimension: 16,
+                child: CircularProgressIndicator(
+                  strokeWidth: 2,
+                  color: ShadTheme.of(context).colorScheme.primaryForeground,
+                ),
+              ),
               child: const Text('Please wait'),
             ),
           PlagroundButtonStyle.gradientShadow => ShadButton(